requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
openai==1.76.0
requests==2.32.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9